+++ R/cut_sels.R
@@ -6,7 +6,6 @@
 #' labels = c("sound.files", "selec"), overwrite = FALSE, norm = FALSE, ...)
 #' @param X object of class 'selection_table', 'extended_selection_table' or data frame containing columns for sound file name (sound.files), 
 #' selection number (selec), and start and end time of signals (start and end).
-#' The output of \code{\link{auto_detec}} can be used as the input data frame. 
 #' @param mar Numeric vector of length 1. Specifies the margins adjacent to the start and end points of selections,
 #' delineating spectrogram limits. Default is 0.05.
 #' @param parallel Numeric. Controls whether parallel computing is applied.
@@ -22,7 +21,7 @@
 #' @param overwrite Logical. If \code{TRUE} sound files with the same name will be 
 #' overwritten. Default is \code{FALSE}.
 #' @param norm Logical indicating whether wave objects must be normalized first using the function \code{\link[tuneR]{normalize}}. Additional arguments can be passed to \code{\link[tuneR]{normalize}} using `...`.` Default is \code{FALSE}. See \code{\link[tuneR]{normalize}} for available options.
-#' @param ... Additional arguments to be passed to the internal \code{\link[tuneR]{normalize}}  function for customizing sound file output. Ignored if  \code{norm = FALSE}. 
+#' @param ... Additional arguments to be passed to the internal \code{\link[tuneR]{normalize}} function for customizing sound file output. Ignored if  \code{norm = FALSE}. 
 #' @return Sound files of the signals listed in the input data frame.
 #' @family selection manipulation
 #' @seealso \code{\link{seltailor}} for tailoring selections